+++ controllers/ValidasikualifikasipenyediaController.php
@@ -417,7 +417,12 @@ class ValidasikualifikasipenyediaController extends Controller {
     }
     public function actionDelete($id) {
         $request = Yii::$app->request;
-        $this->findModel($id)->delete();
+        $model=$this->findModel($id);
+        if($model->paketpengadaan->pemenang){
+                Yii::$app->session->setFlash('warning', 'PaketPengadaan Sudah ada Pemenang');
+                return $this->redirect('index');
+            }
+        $model->delete();
         if ($request->isAjax) {
             Yii::$app->response->format = Response::FORMAT_JSON;
             return ['forceClose' => true, 'forceReload' => '#crud-datatable-pjax'];
@@ -430,6 +435,10 @@ class ValidasikualifikasipenyediaController extends Controller {
         $pks = explode(',', $request->post('pks'));
         foreach ($pks as $pk) {
             $model = $this->findModel($pk);
+            if($model->paketpengadaan->pemenang){
+                Yii::$app->session->setFlash('warning', 'PaketPengadaan Sudah ada Pemenang');
+                return $this->redirect('index');
+            }
             $model->delete();
         }
         if ($request->isAjax) {
